item list component for inbox and today and this week
