+++ apps/frontend/src/components/Inbox/InboxItems.tsx
@@ -2,16 +2,11 @@
 
 import React, { useEffect, useCallback, useState } from "react"
 
-import { CalendarIcon, MoveIcon, GithubIcon, MailsIcon } from "lucide-react"
-import Image from "next/image"
-
 import { RescheduleCalendar } from "./RescheduleCalendar/RescheduleCalendar"
-import BoxIcon from "@/public/icons/box.svg"
-import LinearIcon from "@/public/icons/linear.svg"
+import { ItemList } from "@/src/components/atoms/ItemList"
 import { useAuth } from "@/src/contexts/AuthContext"
 import { CycleItem } from "@/src/lib/@types/Items/Cycle"
 import { useCycleItemStore } from "@/src/lib/store/cycle.store"
-import classNames from "@/src/utils/classNames"
 import { getWeekDates } from "@/src/utils/datetime"
 
 export const InboxItems: React.FC = () => {
@@ -125,31 +120,6 @@ export const InboxItems: React.FC = () => {
     [updateItem, session]
   )
 
-  const getSourceIcon = (source: string) => {
-    switch (source) {
-      case "gmail":
-        return <MailsIcon size={14} />
-      case "githubIssue":
-      case "githubPullRequest":
-        return <GithubIcon size={14} />
-      case "linear":
-        return (
-          <Image
-            src={LinearIcon}
-            alt="linear icon"
-            width={14}
-            height={14}
-            className="opacity-50"
-          />
-        )
-      case "march":
-      case "marchClipper":
-        return null
-      default:
-        return null
-    }
-  }
-
   const filteredItems = items.filter((item) => item.status !== "done")
 
   const handleRescheduleCalendar = (
@@ -186,52 +156,12 @@ export const InboxItems: React.FC = () => {
             </div>
           )}
           <div className="flex flex-col gap-2.5">
-            {filteredItems.map((item) => (
-              <button
-                key={item._id}
-                className="hover-text group flex items-start gap-2 py-1 text-primary-foreground outline-none hover:text-foreground focus:text-foreground"
-                onClick={() => handleExpand(item)}
-                data-item-id={item._id}
-              >
-                <div className="flex items-start gap-2 truncate">
-                  <Image
-                    src={BoxIcon}
-                    alt="checkbox icon"
-                    width={12}
-                    height={12}
-                    onClick={(e) => handleDone(e, item._id, item.status)}
-                    className="invisible mt-1 opacity-50 hover:opacity-100 group-hover:visible"
-                  />
-                  <span
-                    className={classNames(
-                      "text-left truncate",
-                      item.type === "link" && "group-hover:underline"
-                    )}
-                  >
-                    {item.title}
-                  </span>
-                  {item.source !== "march" && (
-                    <div className="mt-[3px] flex items-center text-secondary-foreground">
-                      {getSourceIcon(item.source)}
-                    </div>
-                  )}
-                </div>
-                <div className="invisible mt-[3px] flex items-center gap-2 text-secondary-foreground group-hover:visible">
-                  <CalendarIcon
-                    size={14}
-                    className="hover-text"
-                    onClick={(e) =>
-                      handleRescheduleCalendar(e, item._id, item.dueDate)
-                    }
-                  />
-                  <MoveIcon
-                    size={14}
-                    className="hover-text"
-                    onClick={(e) => e.stopPropagation()}
-                  />
-                </div>
-              </button>
-            ))}
+            <ItemList
+              items={filteredItems}
+              handleExpand={handleExpand}
+              handleDone={handleDone}
+              handleRescheduleCalendar={handleRescheduleCalendar}
+            />
           </div>
         </div>
       )}
